% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellviz3d-package.R
\docType{package}
\name{cellviz3d-package}
\alias{cellviz3d-package}
\alias{cellviz3d}
\title{cellviz3d-package}
\description{
Propose visualization solutions for meshes and points with Plotly and Shiny.
}
\section{Visualisation of meshes}{


See the functions \code{\link{plotlyMesh}}, \code{\link{plotlyMesh_all}}, \code{\link{show_tensors}} and \code{\link{show_cellcontour}}.
}

\author{
Marion <marion.louveaux@gmail.com>
}
