% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alltraces.R
\name{create_alltraces}
\alias{create_alltraces}
\title{Create all traces for a time serie}
\usage{
create_alltraces(meshExample, graphExample, meshColors, display,
  defaultColor = list("#CCCCFF", 0.2), heatmapParam = "GeometryArea")
}
\arguments{
\item{meshExample}{One or several meshes.}

\item{graphExample}{One or several cell graphs.}

\item{meshColors}{Colors for each mesh.}

\item{display}{Display type. 'labels', 'heatmap' or 'none'.}

\item{defaultColor}{Uniform mesh color (HEX color and alpha value between 0 and 1). Needed only if 'none' display type is chosen.}

\item{heatmapParam}{Variable (from cell graph) to display if 'heatmap' display type is selected.}
}
\value{
A list.
}
\description{
Create all traces for a time serie
}
