% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coords.R
\name{find_coords}
\alias{find_coords}
\title{Find coordinates of a tensor}
\usage{
find_coords(cellcenter, tipUnit, d, scale = 1)
}
\arguments{
\item{cellcenter}{x, y, z coordinates.}

\item{tipUnit}{x, y, z coordinates.}

\item{d}{Norm of the vector.}

\item{scale}{Shorten or lengthen the line for visualisation.}
}
\value{
a tibble with coordinates (x1, y1, z1, x2, y2, z2)
}
\description{
Find coordinates of a tensor
}
