% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2hours.R
\name{int2hours}
\alias{int2hours}
\title{Convert an integer to hour minutes in digital format
Assumes that integers (0, 1, 2) refers to 30 minutes timesteps (00:00, 00:30, 01:00)...}
\usage{
int2hours(x)
}
\arguments{
\item{x}{integer}
}
\value{
a string with the hour in digital format (HH:mm)
}
\description{
Convert an integer to hour minutes in digital format
Assumes that integers (0, 1, 2) refers to 30 minutes timesteps (00:00, 00:30, 01:00)...
}
