% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyMesh_all.R
\name{plotlyMesh_all}
\alias{plotlyMesh_all}
\title{A layout for plotly graphs with a time slider}
\usage{
plotlyMesh_all(meshExample, graphExample, meshColors, display,
  defaultColor = list("#CCCCFF", 0.2), heatmapParam)
}
\arguments{
\item{meshExample}{One or several meshes.}

\item{graphExample}{One or several cell graphs.}

\item{meshColors}{Colors for each mesh.}

\item{display}{Display type. 'labels', 'heatmap' or 'none'.}

\item{defaultColor}{Uniform mesh color (HEX color and alpha value between 0 and 1). Needed only if 'none' display type is chosen.}

\item{heatmapParam}{Variable (from cell graph) to display if 'heatmap' display type is selected.}
}
\value{
plotly graph
}
\description{
This function provides a basic layout for plotly graphs with a time sliders.
It can be used to display MGX meshes of a time serie.
}
\examples{
\dontrun{
mesh.all <- readRDS(system.file("extdata",
 "mgx/2013-02-12_LTi6B_dis_A_all_normalMesh_full.rds",
  package = "cellviz3d"))
cellgraph.all <- readRDS(system.file("extdata",
 "mgx/2013-02-12_LTi6B_dis_A_all_cellGraph_full.rds",
  package = "cellviz3d"))

  # Heatmap of cell area for all timepoints of the dataset
  plotlyMesh_all <- function(meshExample = mesh.all,
   graphExample = cellgraph.all,
    meshColors = list(NULL, NULL, NULL, NULL, NULL),
    display = 'heatmap',
    defaultColor = list("#CCCCFF", 0.2),
    heatmapParam = "GeometryArea")

}
}
