% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlySpots.R
\name{plotlySpots}
\alias{plotlySpots}
\title{A layout for 3D scatter plotly graphs for a single timepoint}
\usage{
plotlySpots(Spots_df, x, y, z, timepoint, singleTimepoint, number, color)
}
\arguments{
\item{Spots_df}{dataframe containing x,y,z coordinates, timepoints and number of spots per tiempoint (for instance generated with Spots.as.dataframe)}

\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}

\item{timepoint}{column of the timepoints}

\item{singleTimepoint}{single timepoint to display}

\item{number}{column with the number of spots per timepoint}

\item{color}{continuous value used for coloring the spots}
}
\description{
A layout for 3D scatter plotly graphs for a single timepoint
}
