% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlySpots_all.R
\name{plotlySpots_all}
\alias{plotlySpots_all}
\title{A layout for 3D scatter plotly graphs with a time player  - Many thanks to @friep for her help on this function.}
\usage{
plotlySpots_all(Spots_df, x, y, z, timepoint, number, color)
}
\arguments{
\item{Spots_df}{dataframe containing x,y,z coordinates, timepoints and number of spots per timepoint (for instance generated with Spots.as.dataframe)}

\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}

\item{timepoint}{column of the timepoints}

\item{number}{column with the number of spots per timepoint}

\item{color}{continuous value used for coloring the spots}
}
\description{
A layout for 3D scatter plotly graphs with a time player  - Many thanks to @friep for her help on this function.
}
