% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cellcontour.R
\name{show_cellcontour}
\alias{show_cellcontour}
\title{Add cell contour visualisation to a plotly graph of type mesh3d}
\usage{
show_cellcontour(p, mesh)
}
\arguments{
\item{p}{A plotly graph created with plotlyMesh().}

\item{mesh}{The mesh used to create this plotly graph.}
}
\value{
a plotly graph
}
\description{
Add cell contour visualisation to a plotly graph of type mesh3d
}
\examples{
\dontrun{
myMesh <- readRDS(system.file("extdata",
 "mgx/2013-02-12_LTi6B_dis_A_T0_cells_minW1_normalMesh_full.rds",
  package = "cellviz3d"))
myCellGraph <- readRDS(system.file("extdata",
 "mgx/2013-02-12_LTi6B_dis_A_T0_cells_minW1_cellGraph_full.rds",
  package = "cellviz3d"))

 # With a uniform mesh color:
p <- plotlyMesh(meshExample = myMesh,
meshCellcenter = meshCellcenter) \%>\%
layout(scene =
list(aspectmode = "data"))

show_cellcontour(p, myMesh)
}
}
